Extract shared JUnit XML parser used by gotest and nunit runners
Replaces the two duplicated runner-specific parsers with a single
loadAndParseJUnitXML function in junit_xml.go.
